docs(root): expand on testing guidance
Added testing documentiation from a test app, including codePreview component to show code examples, and a testing component to show how to test components with Jest, RTL and Cypress.
Also added a testing stackblitz button to the testing documentation to allow users to test the code in a stackblitz environment.